+++ lib/src/list_extensions.dart
@@ -5,4 +5,10 @@ extension ListExtensions<E> on List<E> {
   /// If current list is `null` - new list with [element] will be created.
   List<E> copyWith(E element) =>
       this == null ? [element] : (List.from(this)..add(element));
+
+  /// Copy current list with adding all [elements] at the end of new list.
+  ///
+  /// If current list is `null` - copy of list [elements] will be created.
+  List<E> copyWithAll(List<E> elements) =>
+      this == null ? List.from(elements) : (List.from(this)..addAll(elements));
 }